broke app.js into a runner and a set of exports, so that we can either run as node_modules/.bin/flashlight, or as a gulp worker.
