+++ package.json
@@ -4,14 +4,22 @@
   "description": "A search utility for Firebase",
   "keywords": ["firebase", "elasticsearch"],
   "homepage": "https://github.com/firebase/flashlight",
-  "author": {
-    "name": "Michael \"Kato\" Wulf",
-    "email": "katowulf@gmail.com"
-  },
+  "license": "MIT",
+  "contributors": [
+    {
+      "name": "Michael \"Kato\" Wulf",
+      "email": "katowulf@gmail.com"
+    },
+    {
+      "name": "Richard Downe",
+      "email": "richard@casetext.com"
+    }
+  ],
   "engines": {
     "node": "0.10.x"
   },
-  "main": "app.js",
+  "main": "module.js",
+  "bin": { "flashlight": "./app.js" },
   "devDependencies": {},
   "dependencies": {
     "bunyan": "^1.2.3",
